chore: rename client/client to client_api. other clippy fixes
