+++ src/client/client_api/blob_storage.rs
