+++ src/client/client_api/register_apis.rs
